Roles and permission create,update and delete done at object level. Test cases added.
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@10805 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ mifos/src/org/mifos/application/productdefinition/business/LoanOfferingBO.java
@@ -416,11 +416,11 @@ public class LoanOfferingBO extends PrdOfferingBO {
 		if (intDedDisbursement) {
 			this.gracePeriodType = new GracePeriodTypeEntity(
 					GraceTypeConstants.NONE);
-			this.gracePeriodDuration = 0;
+			this.gracePeriodDuration = (short)0;
 		} else if (gracePeriodType.getId().equals(
 				GraceTypeConstants.NONE.getValue())) {
 			this.gracePeriodType = gracePeriodType;
-			this.gracePeriodDuration = 0;
+			this.gracePeriodDuration = (short)0;
 		} else {
 			this.gracePeriodType = gracePeriodType;
 			this.gracePeriodDuration = gracePeriodDuration;
